requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic==2.5.3

